Update Assignment Week 3.R
The assignmment consisted of creating forth functions. The first creates a numeric vector. 
The second function gets the mean of an input vector. If that vector is not available it defaults to the data stored in the cache.
The third function creates a matrix vector.
The final function creates the inverse matrix of a matrix that is given as input. If no correct input is available it defaults to the cache